% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets}
\alias{datasets}
\title{datasets wrapper}
\usage{
datasets(do = c("summary", "download"), args, file = NULL)
}
\arguments{
\item{do}{"summary" or "download"}

\item{args}{all other arguments as a vector}

\item{file}{file path to save the output if desired}
}
\description{
This is a simple wrapped for the datasets command line tool.
}
